// frontend/src/api/generated/models/sheetFormatList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */
import type { SheetFormat } from "./sheetFormat";
import type { SheetFormatOption } from "./sheetFormatOption";

export interface SheetFormatList {
  options: SheetFormatOption[];
  items: SheetFormat[];
}